fix: simplify async_conversation_run example to reduce CI flakiness
The original example used code_interpreter with differential equations,
which caused timeouts and flaky CI failures. Simplified to "2+2" math.

Original complex example preserved as async_conversation_run_code_interpreter.py
and added to CI skip list (too slow/flaky for CI).
